BabelMailSend/requirements.txt
googletrans==4.0.0-rc1
kafka-python